update kvstore client and server with test
